requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
asgiref==3.11.1
attrs==26.1.0
bitarray==3.8.1
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1749686179973/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1764017066538/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1762976168352/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1761202954785/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
ckzg==2.1.7
cloudinary==1.44.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1764077581417/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1764081326783/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
cytoolz==1.1.0
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1763082862281/work
frozenlist==1.8.0
gunicorn==25.3.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hexbytes==1.3.1
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
inflection==0.5.1
iniconfig==2.3.0
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1756754242108/work
libmambapy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_libmambapy_1764158595/work/libmambapy
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1761299902298/work
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1762504092552/work
multidict==6.7.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
parsimonious==0.10.0
pillow==12.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pluggy_1764896838/work
propcache==0.4.1
psycopg2-binary==2.9.12
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1757744645636/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pycryptodome==3.23.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest==9.0.3
pytest-django==4.12.0
python-decouple==3.8
pytz==2026.1.post1
pyunormalize==17.0.0
PyYAML==6.0.3
qrcode==8.2
regex==2026.4.4
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rlp==4.1.0
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1761160813894/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1760564299908/work
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
toolz==1.1.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1753886790/work
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
web3==7.15.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.12.0
yarl==1.23.0
zstandard==0.25.0
